requirements.txt
discord.py>=2.4.0
openai>=1.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
PyYAML>=6.0
aiosqlite>=0.19.0
scipy>=1.11.0
pandas>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
alpaca-py>=0.30.0
fastapi>=0.100.0
uvicorn>=0.23.0
